Moving To Direct Write System
Starting to move to a direct write system. Previously telemetry events where saved in an array and written to CSV at the end of runtime. However, in the event of a crash - they would not be saved.

So now events will be written straight to the CSV. This ensures the data is saved and this also reduces memory consumption as events are cleared as soon as they are written.

Removed FToonTelemetryEvent.